Working now with NFS mount
Got vagrant NFS share working with Gogs data directory
and Docker Compose brings up the services.

Used NFS Manager to get the export on the macOS side
working.

The biggest problem was that the the vagrant VM was
timing out when using UDP for the NFS mount.  Used
the nfs_udp: false option in the mount to default to
TCP.